Use IProvisioningAgent.getService(Class) method
* Returns the correct type so no casting needed
* Some projects were not flagging unnecessary cast by default so changed
their project settings
* Fix few additional unnecessary casts.

+++ tycho-bundles/org.eclipse.tycho.p2.resolver.impl/src/main/java/org/eclipse/tycho/p2/impl/publisher/SourcesBundleDependencyMetadataGenerator.java
@@ -104,7 +104,7 @@ public class SourcesBundleDependencyMetadataGenerator extends AbstractMetadataGe
     }
 
     public long createId(String sourceBundleSymbolicName, String version) {
-        return (long) sourceBundleSymbolicName.hashCode() | (((long) version.hashCode()) << 32);
+        return sourceBundleSymbolicName.hashCode() | (((long) version.hashCode()) << 32);
     }
 
 }
